+++ packages/ayazmo-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ayazmo
 
+## 0.6.8
+
+### Patch Changes
+
+- fix public plugins path
+
+- Updated dependencies []:
+  - @ayazmo/utils@0.6.8
+  - @ayazmo/core@0.6.8
+  - @ayazmo/types@0.6.8
+
 ## 0.6.7
 
 ### Patch Changes