+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ayazmo/types
 
+## 0.6.8
+
+### Patch Changes
+
+- fix public plugins path
+
 ## 0.6.7
 
 ### Patch Changes